+++ pokerkit/analysis.py
@@ -226,6 +226,33 @@ def calculate_equities(
 
     >>> from concurrent.futures import ProcessPoolExecutor
     >>> from pokerkit import *
+    >>> calculate_equities(
+    ...     (
+    ...         parse_range('33'),
+    ...         parse_range('33'),
+    ...     ),
+    ...     Card.parse('Tc8d6h4s'),
+    ...     2,
+    ...     5,
+    ...     Deck.STANDARD,
+    ...     (StandardHighHand,),
+    ...     sample_count=1000,
+    ... )
+    [0.5, 0.5]
+    >>> calculate_equities(
+    ...     (
+    ...         parse_range('2h2c'),
+    ...         parse_range('3h3c'),
+    ...         parse_range('AhKh'),
+    ...     ),
+    ...     Card.parse('3s3d4c'),
+    ...     2,
+    ...     5,
+    ...     Deck.STANDARD,
+    ...     (StandardHighHand,),
+    ...     sample_count=1000,
+    ... )
+    [0.0, 1.0, 0.0]
     >>> with ProcessPoolExecutor() as executor:
     ...     calculate_equities(
     ...         (
@@ -243,33 +270,6 @@ def calculate_equities(
     ...     )
     ...
     [0.0, 0.0, 1.0]
-    >>> calculate_equities(
-    ...     (
-    ...         parse_range('2h2c'),
-    ...         parse_range('3h3c'),
-    ...         parse_range('AhKh'),
-    ...     ),
-    ...     Card.parse('3s3d4c'),
-    ...     2,
-    ...     5,
-    ...     Deck.STANDARD,
-    ...     (StandardHighHand,),
-    ...     sample_count=1000,
-    ... )
-    [0.0, 1.0, 0.0]
-    >>> calculate_equities(
-    ...     (
-    ...         parse_range('3d3h'),
-    ...         parse_range('3c3s'),
-    ...     ),
-    ...     Card.parse('Tc8d6h4s'),
-    ...     2,
-    ...     5,
-    ...     Deck.STANDARD,
-    ...     (StandardHighHand,),
-    ...     sample_count=1000,
-    ... )
-    [0.5, 0.5]
 
     :param hole_ranges: The ranges of each player in the pot.
     :param board_cards: The board cards, may be empty.
@@ -321,6 +321,90 @@ def calculate_equities(
         equities[i] = equity / sample_count
 
     return equities
+
+
+def calculate_hand_strength(
+        player_count: int,
+        hole_range: Iterable[Iterable[Card]],
+        board_cards: Iterable[Card],
+        hole_dealing_count: int,
+        board_dealing_count: int,
+        deck: Deck,
+        hand_types: Iterable[type[Hand]],
+        *,
+        sample_count: int,
+        executor: Executor | None = None,
+) -> float:
+    """Calculate the hand strength: odds of beating a single other hand
+    chosen uniformly at random.
+
+    The user may supply an executor to use parallelization. If not
+    given, a single-threaded evaluation is performed.
+
+    >>> from concurrent.futures import ProcessPoolExecutor
+    >>> from pokerkit import *
+    >>> calculate_hand_strength(
+    ...     3,
+    ...     parse_range('3h3c'),
+    ...     Card.parse('3s3d2c'),
+    ...     2,
+    ...     5,
+    ...     Deck.STANDARD,
+    ...     (StandardHighHand,),
+    ...     sample_count=1000,
+    ... )
+    1.0
+    >>> with ProcessPoolExecutor() as executor:
+    ...     calculate_hand_strength(
+    ...         3,
+    ...         parse_range('AsKs'),
+    ...         Card.parse('QsJsTs'),
+    ...         2,
+    ...         5,
+    ...         Deck.STANDARD,
+    ...         (StandardHighHand,),
+    ...         sample_count=1000,
+    ...         executor=executor,
+    ...     )
+    ...
+    1.0
+
+    :param player_count: Number of players in the pot.
+    :param hole_range: The range of the player.
+    :param board_cards: The board cards, may be empty.
+    :param hole_dealing_count: The final number of hole cards; for
+                               hold'em, it is ``2``.
+    :param board_dealing_count: The final number of board cards; for
+                                hold'em, it is ``5``.
+    :param deck: The deck; most games typically use
+                 :attr:`pokerkit.utilities.Deck.STANDARD`.
+    :param hand_types: The hand types; most games typically just use
+                       :class:`pokerkit.hands.StandardHighHand`.
+    :param sample_count: The number of samples to simulate, higher value
+                         gives greater accuracy and fidelity.
+    :param executor: The optional executor, defaults to ``None`` which
+                     is just using 1 thread/process. The user can supply
+                     a ``ProcessPoolExecutor`` to use processes.
+    :return: The equity values.
+    """
+    hole_ranges: list[Iterable[Iterable[Card]]] = [
+        [[]] for _ in range(player_count - 1)
+    ]
+
+    hole_ranges.append(hole_range)
+
+    equities = calculate_equities(
+        hole_ranges,
+        board_cards,
+        hole_dealing_count,
+        board_dealing_count,
+        deck,
+        hand_types,
+        sample_count=sample_count,
+        executor=executor,
+    )
+
+    return equities[-1]
 
 
 @dataclass
@@ -382,7 +466,7 @@ class Statistics:
         )
 
     @property
-    def sample_size(self) -> int:
+    def sample_count(self) -> int:
         """Return the sample size.
 
         :return: The sample size.